requirements.txt
anyio==3.5.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
blinker==1.4
bs4==0.0.1
catfish==1.4.13
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.9
Click==7.0
colorama==0.4.3
command-not-found==0.3
configparser==5.1.0
crayons==0.4.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==4.0.1 
django-cors-headers==3.10.1
django-crontab==0.7.1
django-elasticsearch-dsl==7.2.1
django-elasticsearch-dsl-drf==0.22.2
django-nine==0.2.5
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.1.0
elasticsearch==7.16.2
elasticsearch-dsl==7.4.0
entrypoints==0.3
fastapi==0.71.0
filelock==3.0.12
gunicorn==20.1.0
h11==0.12.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
lxml==4.7.1
menulibre==2.2.1
more-itertools==4.2.0
mugshot==0.4.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
outcome==1.1.0
pbr==5.8.0
pexpect==4.6.0
Pillow==8.2.0
pipenv==11.9.0
psutil==5.5.1
psycopg2-binary==2.8.6
pycairo==1.16.2
pycodestyle==2.7.0
pycups==1.9.73
pydantic==1.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.1
PyNaCl==1.3.0
pyOpenSSL==21.0.0
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-docx==0.8.11
pytz==2021.1
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.1.0
sgt-launcher==0.2.5
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.2.4
starlette==0.17.1
systemd-python==234
testresources==2.0.1
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.1
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25
uvicorn==0.16.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
webdriver-manager==3.5.2
whitenoise==5.2.0
wsproto==1.0.0
xcffib==0.8.1
xkit==0.0.0
xlwt==1.3.0
zipp==1.0.0
